disabled curl output buffering for grep check
from man curl: DN: -N, --no-buffer: Disables  the  buffering  of the output stream.
In normal work situations, curl will use a standard buffered output stream that
will have the effect that it will output the data in chunks, not necessarily
exactly when the data arrives.  Using this option will disable that buffering.

This way, grep is not closing the pipe while curl stil has chunks left - which
results in a "(23) Failed writing body" error.